--- a/lalrpop/src/lr1/build/mod.rs
+++ b/lalrpop/src/lr1/build/mod.rs
@@ -123,7 +123,9 @@ impl<'grammar, L: LookaheadBuild> LR<'grammar, L> {
                             index,
                             lookahead,
                         },
-                    )| { (symbol, (Item::lr0(production, index), lookahead)) },
+                    )| {
+                        (symbol.clone(), (Item::lr0(production, index), lookahead))
+                    },
                 )
                 .collect();