Running [samply](https://github.com/mstange/samply) to profile `lalrpop`, I found that most of the time spent during the `LR(1) state construction (lane)` phase is in the `continue_trace` recursive function, which calls `can_shift_nonterminal` in a loop. This last function looks like it should just perform a simple equality on two symbol references, but instead unnecessarily clones a symbol.
Other changes include reducing allocations and minor clippy fixes throughout the entire crate.

I've measured both [Solang] and [Rust Python], which yielded similar times, by using the time printed with the `--level verbose` flag:

-   debug: ~7.9s -> ~6.8s
-   release: ~1.1s -> ~0.45s

[solang]: https://github.com/hyperledger/solang/blob/a14139e916a6a89875d4a1e15c9548d0038b24cd/solang-parser/src/solidity.lalrpop
[rust python]: https://github.com/RustPython/RustPython/blob/c9546c2419f2e4db7e2987f30f0508dc29b30035/compiler/parser/python.lalrpop#L1

Another similar performance improvement can be seen by switching the `Map` and `Set` types to the standard `HashMap` with the [`Fx` hasher][fx], but I ran into a lot of test failures so I didn't include it in this PR. If this change is implemented correctly, I believe it can save another 1-2s on the debug build and ~0.1s on the release build.

[fx]: https://github.com/rust-lang/rustc-hash