@youknowone 2023-04-17
Could you split the PR to trivial fix and actual speed up? The diff is large so it will be helpful to review.