@DaniPopes 2023-04-18
Hey, thanks for the comment! The diff seems large at first glance but if you enable "Hide whitespace" on the Github settings and review commit-by-commit it will make it much easier.

The main performance change is in 187833a, Let me know if you still want me to create a new PR for the other changes.