@yannham 2023-04-18
@DaniPopes thanks a lot for the PR, the numbers do look good.

I'll take a deeper look at it as well, but out of curiosity, what was your rationale for the `#[inline]` annotations? My understanding is that inlining is a complex matter, and most of the time `rustc` (well, actually, LLVM) probably knows better. The situation where it can be interesting to do so is:

- for libraries (as, as far as I understand, there's no cross-crate inlining happening without such an annotation)
- and, inside libraries, for functions that are performance-criticial (because there's a cost to inlining, in compilation time, but also because it can become a pessimisation)

Here I don't think we need cross-crate inlining, so I wonder if you just did a big change and measured that the result was positive, or if you measured specifically that some functions did gain something from inlining?