@DaniPopes 2023-04-18
I saw that the `can_shift_nonterminal` small function accounted for more than half the time spent in the release build, so I preemptively added the `#[inline]` attribute, but it looks like the compiler/LLVM now can inline it by itself after removing the expensive `clone`. It's about 10% slower with a manual `#[inline(never)]`, so I think it's better to keep it.