@youknowone 2023-04-18
@yannham unlike `#[inline(never)]` or `#[inline(always)]`, `#[inline]` is not that strong. 

By looking the function size, most of them looks reasonable. Not sure about comparably bigger ones.