@yannham 2023-04-19
> I saw that the can_shift_nonterminal small function accounted for more than half the time spent in the release build, so I preemptively added the #[inline] attribute, but it looks like the compiler/LLVM now can inline it by itself after removing the expensive clone. It's about 10% slower with a manual #[inline(never)], so I think it's better to keep it.

I think if this doesn't make a difference with no inline annotation at all in your working example (of course `#[inline(never)]` is going to be worse, but I think that's a bit beside the point), I would personally strip off the annotations. I know `#[inline]` isn't too strong, but it's still overriding in some way the decisions of `rustc` and `LLVM`, and to the best of my knowledge, isn't a common practice when you don't care about cross-crate inlining. Also, when you start to add inline hints, you have to start thinking well why this function and not this function, and you either have to come with a rationale or make somehow arbitrary micro-decisions. If we can just forget about inlining altogether and let that work to the optimizer, I looks like a win in the end!

I general I would vouch for adding code/changing code for performance reason when it's clearly motivated, that is, if adding the `#[inline]` does improve the situation in a measurable way with no annotation at all, and not as a kind of preventive measure. This just my humble opinion, and I won't die on this hill either if you're absolutely sure that this is a better idea to keep the annotation :slightly_smiling_face: 

Maybe @nikomatsakis has something to say here as well, as he's probably more competent than me on those kind of matters.