@yannham 2023-04-21
@DaniPopes I started to review but there are spurious diff due to a PR implementing clippy suggestions that has been merged in between. Could you try a rebase please whenever you have the time to do so, and then ping me?