@youknowone 2023-04-21
@yannham I rebased it for review if you have time now.

@DaniPopes Please rebase it again to verify signing when you have time. And additional notes. Me and @yannham are not the original owners of this repository, so tending to being conservative on decisions when we don't know about consensus. (@yannham do you agree?)

I think this PR will be more easily merged by being split to 5 pieces 
1. clippy before 1.56 (2021 edition)
2. clippy after 1.56
3. RustWrite
4. removing allocations
5. adding `#[inline]`. 

1, 3, 4 looks to be merged quickly. 2 has dependency. no idea about 5.