@DaniPopes 2023-04-21
@yannham @yannham I've rebased onto master and squashed the commits down to one, since they have been split off into different PRs (points 1 and 2: #745, point 3: #744). I've removed all `#[inline]` attributes (point 5), which leaves this PR addressing point number 4.