@youknowone 2023-04-22
@DaniPopes Thank you from RustPython by using it for benchmark 😉